requirements.txt
openai>=1.66.0
langchain-community==0.3.19
python-dotenv==0.21.0
httpx==0.28.1
pydantic==2.10.4
uvicorn==0.34.0
fastapi==0.115.14
deepeval>=1.0.0
groq>=0.9.0
pytest==7.4.0
pytest-asyncio==0.20.3
faiss-cpu==1.8.0.post1
langchain-text-splitters==0.3.11
setuptools<81
tensorflow==2.16.2
tensorflow-hub==0.16.1